// src/types/react.grid.line.properties.type.d.ts
declare type ReactGridDrawLineOptionalProperties = {
    lineClickTolerance?: number;
    selectCircleSize?: number;
    circleLineShiftSize?: number;
    contextLineWidth?: number;
    lineColour?: string;
};
declare type ReactGridDrawLineRequiredProperties = {
    lineClickTolerance: number;
    selectCircleSize: number;
    circleLineShiftSize: number;
    contextLineWidth: number;
    lineColour: string;
};